Set minimum criteria for system package
There are some features in 1.5.X that I think we'll want to leverage.

Also work around what I believe is a packaging bug in Ubuntu Eoan and
newer. I'm not 100% certain, but I believe that libbenchmark_main
doesn't work correctly when it's not a shared library.

This means that right now, none of the supported platforms have an
acceptable version of `benchmark` available as a system package.
However, Fedora's package seems to work OK and I was able to confirm
that this package behaves correctly in the presence of the system
package.

Signed-off-by: Scott K Logan <logans@cottsay.net>