Set the SOVERSION on benchmark libraries
The built-in logic for detecting the SOVERSION on the shared libraries
requires that the sources are in a tagged git repository, but we're
downloading a tarball. If we disable the git tag extraction and manually
specify the version that we downloaded, the SOVERSION is correct.

Both Ubuntu and Fedora benchmark packages work around the SOVERSION in
this way.

Signed-off-by: Scott K Logan <logans@cottsay.net>